dana/requirements.txt
flake8
openai
psycopg2-binary
pytest
pytest-mock
python-dotenv
pyyaml
sqlalchemy
pgvector